% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gibbs_dp}
\alias{gibbs_dp}
\title{Collapsed Gibbs sampler for infinite Bernoulli Mixture Model}
\usage{
gibbs_dp(data, nsamples, alpha = NULL, a = 1, b = 1, beta = 0.5,
  gamma = 0.5, burnin = NULL, relabel = FALSE, burnrelabel = 50,
  maxK = 30, debug = FALSE)
}
\arguments{
\item{data}{Data frame or matrix with observations in rows and binary
variables in columns}

\item{nsamples}{Number of samples to take}

\item{alpha}{Starting value of alpha}

\item{a}{First Gamma parameter on prior on Alpha}

\item{b}{Second Gamma parameter on prior on Alpha}

\item{beta}{First Beta parameter on prior used for all Bernoulli
variables across all components}

\item{gamma}{Second Beta parameter on prior used for all Bernoulli
variables across all components}

\item{burnin}{Number of samples to discard at start of chain}

\item{relabel}{Whether to apply Stephens 2000b relabelling method}

\item{burnrelabel}{If \code{relabel} is set then this parameter controls
how many samples (from the end of the burnin period) are used for
the initial batch run of Stephens' algorithm. Following this, the 
online implementation is applied at each sample.}

\item{maxK}{Maximum number of clusters to assume. Reduces computational storage
requirements.}

\item{debug}{Whether to display debug messages.}
}
\value{
A list with sampled components z, theta, alpha.
  If \code{relabel} is set then it also returns permutations,
  z_relabelled, and theta_relabelled
}
\description{
Samples from the Dirichlet Process prior infinite BMM using
the Chinese Restaurant Process formulation
}
