% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gibbs_full}
\alias{gibbs_full}
\title{Full Gibbs sampler for finite Bernoulli Mixture Model}
\usage{
gibbs_full(data, nsamples, K, alpha = 1, beta = 0.5, gamma = 0.5,
  burnin = NULL, debug = FALSE)
}
\arguments{
\item{data}{Data frame or matrix with observations in rows and binary
variables in columns}

\item{nsamples}{Number of samples to take}

\item{K}{Number of mixtures}

\item{alpha}{Concentration parameter}

\item{beta}{First Beta parameter on prior used for all Bernoulli
variables across all components}

\item{gamma}{Second Beta parameter on prior used for all Bernoulli
variables across all components}

\item{burnin}{Number of samples to discard at start of chain}

\item{debug}{Whether to display debug messages.}
}
\value{
A list with sampled values pi, z, and theta.
}
\description{
Full Gibbs sampler for finite Bernoulli Mixture Model
}
