% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmm-mcmc.R
\docType{data}
\name{K3_N1000_P5}
\alias{K3_N1000_P5}
\title{Simulated mixture dataset.}
\format{A matrix with 1000 rows and 5 columns}
\description{
Has 1000 observations of 5 Bernoulli variables from 3 clusters.
}
\details{
The clusters are split in the ratio 0.6:0.2:0.2.
The true theta Bernoulli parameters are:
[[0.7, 0.8, 0.2, 0.1, 0.1],
 [0.3, 0.5, 0.9, 0.8, 0.6],
 [0.1, 0.2, 0.5, 0.4, 0.9]]
}
