% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_gibbs}
\alias{plot_gibbs}
\title{Plots samples from Bernoulli mixture models}
\usage{
plot_gibbs(obj, theta = TRUE, z = TRUE, pi = FALSE, heights = NULL,
  cluster_threshold = 0.1, cluster_labels = NULL,
  theta_labels = NULL, theta_to_display = NULL)
}
\description{
Plots samples from Bernoulli mixture models
}
